+++ src/babashka/main.clj
@@ -39,6 +39,7 @@
    [babashka.impl.http-client :refer [http-client-namespace
                                       http-client-websocket-namespace
                                       http-client-interceptors-namespace]]
+   [babashka.impl.markdown :as md]
    [babashka.impl.nrepl-server :refer [nrepl-server-namespace]]
    [babashka.impl.pods :as pods]
    [babashka.impl.pprint :refer [pprint-namespace]]
@@ -432,7 +433,9 @@ Use bb run --help to show this help output.
        'babashka.cli cli/cli-namespace
        'babashka.http-client http-client-namespace
        'babashka.http-client.websocket http-client-websocket-namespace
-       'babashka.http-client.interceptors http-client-interceptors-namespace}
+       'babashka.http-client.interceptors http-client-interceptors-namespace
+       'nextjournal.markdown md/markdown-namespace
+       'nextjournal.markdown.utils md/markdown-utils-namespace}
     features/xml? (assoc 'clojure.data.xml @(resolve 'babashka.impl.xml/xml-namespace)
                          'clojure.data.xml.event @(resolve 'babashka.impl.xml/xml-event-namespace)
                          'clojure.data.xml.tree @(resolve 'babashka.impl.xml/xml-tree-namespace))